feature/YAS-14-get-rid-of-recursion-in-filesyste

+++ Model/HistoryStorage.cpp
@@ -2,6 +2,8 @@
 
 void HistoryStorage::Add(const ImportBodyMessage::ImportBodyItem& item)
 {
+    if (item._systemItemType == SystemItemType::FOLDER)
+        return;
     (*this)[item.id] = item;
     if (size() > _max_available_size)
         DeleteOld(item.date_ms);
@@ -11,7 +13,7 @@ void HistoryStorage::DeleteOld(long long date)
 {
     vector<string> old;
     for (const auto& [id, item] : *this)
-        if (item.date_ms < date)
+        if (item.date_ms < date - day_ms)
             old.push_back(id);
     for (const auto& it : old)
         erase(it);
@@ -28,4 +30,10 @@ vector<ImportBodyMessage::ImportBodyItem> HistoryStorage::GetAll() const {
     for (const auto& [id, item]: *this)
         ans.push_back(item);
     return ans;
+}
+
+void HistoryStorage::Remove(const ImportBodyMessage::ImportBodyItem &item) {
+    if (item._systemItemType == SystemItemType::FOLDER)
+        return;
+    Remove(item.id);
 }
